Add files via upload
This project is on the Logistic Regression. The dataset comes from the UCI Machine Learning repository, and it is related to direct marketing campaigns (phone calls) of a Portuguese banking institution. The classification goal is to predict whether the client will subscribe (1/0) to a term deposit (variable y). The dataset provides the bank customers’ information. It includes 41,188 records and 21 fields.